[201911] Fix snmp subagent errors in shutdown path (#259)
Fixed an error raised in snmp container stop path. The issue is seen when SIGTERM is send to SNMP supervisord, and it attepmts to stop snmp-subagent. There is a bug in subagent's shutdown path where below erros is seen upon receiving SIGTERM

This issue turned out to be very costly in warmboot shutdown path. The SIGTERM sent to subagent never shutsdown the agent as it crashed. So supervisord waits for 10s before sending SIGKILL.
In some instances even kill signal to agent fails as the subagent has crashed, this makes dockerd to wait for 10 extra seconds to send SIGKILL to the container itself.
The 20s wait in the shutdown process is not only costly in terms of time spent, but it also makes dockerd to lock resources causing other docker exec commands to timeout.